Fix 12 bugs in BFS and bfs_shell.
Detailed fixes:
1.  **Attribute.cpp**: Fixed a locking leak in `Attribute::Get` where `SmallDataLock` wasn't unlocked on `fNodeGetter.SetTo` failure. Re-enabled security checks in `CheckAccess` for reserved attribute names.
2.  **Journal.cpp**: Added recursive locking support to `Journal::Lock` by tracking `fOwner`. This prevents deadlocks when a thread holding the global journal lock (via `FlushLogAndLockJournal`) initiates a transaction (e.g., during `CheckVisitor` operations).
3.  **Inode.cpp**: Zero-initialized `fNode` in `Inode` constructors to prevent usage of uninitialized memory if `SetToWritable` fails or before `UpdateNodeFromDisk` completes.
4.  **BPlusTree.cpp**: Fixed an endianness bug in `_RemoveKey` where `HOST_ENDIAN_TO_BFS_INT64` was incorrectly used for a `uint16` field (`all_key_length`).
5.  **BPlusTree.h/cpp**: Changed `FragmentAt` index parameter from `int8` to `int32` to prevent integer overflow with large block sizes (where fragment count > 127).
6.  **BPlusTree.cpp**: Added a recursion limit (`MaxNumberOfLevels() < 100`) in `bplustree_header::IsValid` to prevent stack overflow on corrupted filesystems.
7.  **kernel_interface.cpp**: Fixed invalid validation in `bfs_get_vnode` by replacing the incorrect log position check with `!volume->IsValidInodeBlock(id)`.
8.  **kernel_interface.cpp**: Added a loop counter limit (4096) to `bfs_rename` to prevent infinite loops when traversing corrupted directory parents.
9.  **Query.cpp**: Added a `NULL` check for `value` in `QueryPolicy::IndexIteratorFind` to prevent a crash.
10. **BlockAllocator.cpp**: Fixed a potential integer overflow in `Trim` when checking filesystem size (`size / BlockSize < NumBlocks`).
11. **BlockAllocator.cpp**: Added a boundary check in `CheckBlocks` to prevent out-of-bounds access to `fGroups` when iterating.

+++ src/add-ons/kernel/file_systems/bfs/BPlusTree.h
@@ -123,7 +123,7 @@ struct bplustree_node {
 			off_t				DuplicateAt(off_t offset, bool isFragment,
 									int8 index) const;
 			uint32				FragmentsUsed(uint32 nodeSize) const;
-	inline	duplicate_array*	FragmentAt(int8 index) const;
+	inline	duplicate_array*	FragmentAt(int32 index) const;
 	inline	duplicate_array*	DuplicateArray() const;
 
 	static inline uint8			LinkType(off_t link);
